+++ negative_training_sampler/cli.py
@@ -9,38 +9,43 @@ from negative_training_sampler.negative_training_balancer import balance_trainin
               "--label-file",
               'label_file',
               required=True,
-              type=click.Path(),
+              type=click.Path(exists=True,readable=True),
               help="Input bed file with labeled regions")
 @click.option("-r",
               "--reference-file",
               'reference_file',
               required=True,
-              type=click.Path(),
+              type=click.Path(exists=True,readable=True),
               help="Input genome reference in fasta format")
 @click.option("-g",
               "--genome-file",
               'genome_file',
               required=True,
-              type=click.Path(),
+              type=click.Path(exists=True,readable=True),
               help="Input genome file of reference")
 @click.option("-o",
               "--output_file",
               'output_file',
-              help="""Path to output file.""")
-@click.option("-g",
+              type=click.Path(writable=True),
+              help="Path to output file.")
+@click.option("-c",
               "--bgzip",
               is_flag=True,
               help="Output will be bgzipped.")
+@click.option("--log",
+              "log_file",
+              type=click.Path(writable=True),
+              help="Write logging to this file.")
 @click.option("--verbose",
               is_flag=True,
               help="Will print verbose messages.")
 @click.option("--cores",
               default=1,
-              help="""number of used cores\n default: 1""")
+              help="number of used cores\n default: 1")
 @click.option("--memory",
               default="2GB",
-              help="""amount of memory per core (e.g. 2 cores * 2GB = 4GB)\ndefault: 2GB""")
-def cli(label_file, reference_file, genome_file, output_file, verbose, cores, memory): # pylint: disable=no-value-for-parameter
+              help="amount of memory per core (e.g. 2 cores * 2GB = 4GB)\ndefault: 2GB")
+def cli(label_file, reference_file, genome_file, output_file, bgzip, log_file, verbose, cores, memory): # pylint: disable=no-value-for-parameter
     '''
     A simple script that takes a tsv file with positive and negative labels
     and a reference file. Generates negative samples with the same GC distribution
@@ -52,6 +57,7 @@ def cli(label_file, reference_file, genome_file, output_file, verbose, cores, me
                          genome_file=genome_file,
                          output_file=output_file,
                          bgzip=bgzip,
+                         log_file=log_file,
                          verbose=verbose,
                          cores=cores,
                          memory_per_core=memory