+++ CHANGELOG.md
@@ -4,17 +4,19 @@
 
 - use a genome file of the fasta for the contigs by `--genome-file` (see #7) to avoid hard-coded contigs. Now this sampler should work with every species.
 
+- adding a logger using logging. with --log the log file can be written.
+
 ### Issue #6
 
 - reworked CLI
     - input is now connected to the -i flag
     - refrence is now connected to the -r flag
     - added --verbose flag
-    - added -g (--bgzip) flag
+    - added -c (--bgzip) flag
 
 - added bgzip support for -o
-    - output file is bgzipped if -g flag is set
-    - output file is bgzipped if filename ends with `.gz`, even if -g is not set
+    - output file is bgzipped if -c flag is set
+    - output file is bgzipped if filename ends with `.gz`, even if -c is not set
 
 - changed how output works
     - positive and negative labeled entries are now saved in one file